fixes for dash + code cleanups
Since dash does not support \h and \w, use `$(hostname)` for \h and a
function for \w (that does the tilde $HOME prefix.)

Use a real newline in PS1 instead of \n, dash does not support \n, and
enclose it in escape codes.

Save the escape codes \001 and \002 to variables output with printf for
use in PS1 since dash does not support the escapes.

Use %s for string params to all printf invocations instead of putting
them directly into the format string.