fix prompt width and escapes on zsh
Do not use the `\001` and `\002` escape sequences for zsh, it does not
support them, use `%{` and `%}` instead, which serve the same purpose.

When in zsh also use precmd() for the first line, so as to not trigger
any bizarre behavior due to a multi-line prompt.

Also stop using the escapes for terminating unicode glyph, as it is not
zero-width.

This prompt is now fully compatible with zsh.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>